+++ brixterm/main.py
@@ -9,6 +9,9 @@ parser.add_argument(
     "--dev", action="store_true", help="(optional) Run in development mode " "with Arize Phoenix tracing enabled."
 )
 parser.add_argument(
+    "--light_mode", action="store_true", help="(optional) Optimize looks for light mode terminal (dark is default)."
+)
+parser.add_argument(
     "--model", help=f"(optional) Specify GPT model. (default='{DEFAULT_GPT_MODEL}')", default=DEFAULT_GPT_MODEL
 )
 args = parser.parse_args()
@@ -25,12 +28,20 @@ from brixterm.console_printer import ConsolePrinter
 from brixterm.terminal_app import TerminalApp
 
 gpt_model = args.model
+dark_mode = not args.light_mode
+
+color_mode_var = os.getenv("BRIXTERM_COLOR_MODE")
+if color_mode_var == "dark":
+    dark_mode = True
+if color_mode_var == "light":
+    dark_mode = False
+
 if os.getenv("BRIXTERM_MODEL"):
     gpt_model = os.getenv("BRIXTERM_MODEL")
 
 
 def main():
-    printer = ConsolePrinter()
+    printer = ConsolePrinter(dark_mode=dark_mode)
     executor = CommandExecutor()
     cmd_hist = CommandHistory(size=CMD_HIST_SIZE)
     smart_terminal = SmartTerminal(